Rename compatibility module
